refactor: #1 Entity Class Package 변경
